feat: simplified ai_utils tests
